requirement.txt
pandas
numpy
mysql-connector-python
matplotlib
scikit-learn
tensorflow
